+++ src/main/kotlin/net/zelanton/processkit/ProcessGroup.kt
@@ -67,6 +67,7 @@ public class ProcessGroup(
             command.stdinSource,
             command.stdoutCharset,
             command.stderrCharset,
+            command.keepStdinOpen,
         )
     }